Add files via upload
A Google search engine project using JavaScript, HTML, and CSS could involve creating a web page that functions similarly to the Google search engine. Here are some key features that the project could include:

Search box: The search box would be the main input element on the page. Users would be able to type in their search query and click on the "Search" button to initiate the 
search.
-It includes a header with a search bar and links to Gmail and Google Images, as well as a profile icon.
-the search bar, where users can enter their search queries. 
Then, using JavaScript, the program takes the user's input and sends it to the Google search engine, where the results are returned.

-The server processes the search query and returns a list of search results in JSON format.
the function uses the fetch() method to make a request to the Google Custom Search API with 
a specific API key and query parameter based on the value of input_txt
The JavaScript code receives the JSON response and displays the search results on the web page using HTML and CSS.


the output_desc element to the connect HTML string for each item, which displays the search results 
on the page. If there is an error during the fetch request, the function logs an error message to the 
console.


Search results: The search results would be displayed below the search box. The results could be displayed in a list format, with each result showing the page title, URL, and a brief description